refactor(dev): encapsulate MetaMask Pay test rows
